+++ src/main/java/com/sprint/mission/discodeit/auth/handler/LoginSuccessHandler.java
@@ -1,6 +1,7 @@
-package com.sprint.mission.discodeit.auth;
+package com.sprint.mission.discodeit.auth.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sprint.mission.discodeit.auth.service.DiscodeitUserDetails;
 import com.sprint.mission.discodeit.dto.data.UserDto;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
